.history/requirements_20211130120845.txt
absl-py==1.0.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.7.4
aiormq==3.3.1
alembic==1.7.5
altair==4.1.0
appdirs==1.4.4
APScheduler==3.6.3
argcomplete==1.12.3
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
audioread==2.1.9
autopep8==1.5.7
autopy==4.0.0
backcall==0.2.0
base58==2.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
bleach==3.3.0
blinker==1.4
boto3==1.17.109
botocore==1.20.109
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.8
chitra==0.1.2
click==7.1.2
cloudpickle==1.6.0
cmake==3.20.5
colorama==0.4.4
colorclass==2.2.0
coloredlogs==15.0.1
colorhash==1.0.3
cryptography==3.4.7
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.2
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.4
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
ecdsa==0.17.0
efficientnet==1.0.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.63.0
fbmessenger==6.0.0
filelock==3.0.12
Flask==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fuser==1.0.4
future==0.18.2
gast==0.4.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
gunicorn==20.1.0
h11==0.9.0
h5py==3.6.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.2.0
httpx==0.15.4
humanfriendly==9.2
idna==3.3
image-classifiers==1.0.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.2
importlib-resources==5.2.0
imutils==0.5.4
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonify==0.5
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python==2.0.2
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
libclang==12.0.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mattermostwrapper==2.2
mediapipe==0.8.4.2
mistune==0.8.4
mock==4.0.3
MouseInfo==0.1.3
mplfinance==0.12.7a5
multidict==5.1.0
multitasking==0.0.9
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.3.0
numba==0.53.1
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
opencv-contrib-python==4.5.2.52
opencv-python==4.5.3.56
opencv-python-headless==4.5.2.54
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.2.2
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2021.5.29
pooch==1.4.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.1
psycopg2==2.9.2
psycopg2-binary==2.8.6
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.7.3
pydeck==0.6.2
pydot==1.4.2
PyGetWindow==0.0.9
Pygments==2.10.0
PyInquirer==1.0.3
PyJWT==2.0.1
pykwalify==1.8.0
pymongo==3.10.1
PyMsgBox==1.0.9
pynput==1.7.3
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyreadline==2.1
PyRect==0.1.4
pyrouge==0.1.3
pyrsistent==0.17.3
PyScreeze==0.1.27
PySocks==1.7.1
pyTelegramBotAPI==3.8.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-dotenv==0.15.0
python-engineio==4.2.0
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.3.0
python-telegram-bot==13.6
PyTweening==1.0.3
pytz==2021.1
PyWavelets==1.1.1
pywin32==302
pywinpty==1.0.1
PyYAML==5.3.1
pyzmq==22.3.0
questionary==1.9.0
rasa==2.7.0
rasa-sdk==2.7.0
redis==3.5.3
regex==2020.9.27
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
rocketchat-API==1.15.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
sanic==20.12.3
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.11.2
segmentation-models==1.0.1
selenium==3.141.0
Send2Trash==1.5.0
sentry-sdk==1.1.0
Shapely==1.8.0
six==1.16.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==4.0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.2.4
starlette==0.13.6
streamlit==0.81.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.12.1
tensorflow-datasets==4.3.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.10.0
tensorflow-io==0.19.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.1.0
tensorflow-probability==0.11.1
tensorflowjs==3.6.0
termcolor==1.1.0
terminado==0.9.4
terminaltables==3.1.0
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.2.26
toml==0.10.2
toolz==0.11.1
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.1.0
twilio==6.50.1
typeguard==2.12.1
typer==0.4.0
typing-extensions==4.0.0
tzlocal==2.1
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.7
uvicorn==0.13.3
validators==0.18.2
virtualenv==20.4.7
virtualenv-clone==0.5.4
watchdog==2.1.0
watchgod==0.6
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websockets==8.1
Werkzeug==2.0.2
whitenoise==5.3.0
widgetsnbextension==3.5.1
wrapt==1.13.3
xlrd==2.0.1
yarl==1.6.3
yfinance==0.1.55
zipp==3.6.0
